LayoutTests/crypto/subtle/jwk-export-use-values.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting keys with various usages to JWK.");

jsTestIsAsync = true;

var extractable = true;

var aesKeyAsArrayBuffer = Base64URL.parse("jnOw99oOZFLIEPMrgJB55WL46tJSLGt7");
var hmacKeyAsArrayBuffer = Base64URL.parse("ahjkn-_387fgnsibf23qsvahjkn-_387fgnsibf23qs");

function testWithAESCBC(usages, expectedUse)
{
    return crypto.subtle.importKey("raw", aesKeyAsArrayBuffer, "AES-CBC", extractable, usages).then(function(result) {
        return crypto.subtle.exportKey("jwk", result);
    }).then(function(result) {
        jwk = JSON.parse(bytesToASCIIString(result));
        debug(usages + ":");
        shouldBe("jwk.use", "'" + expectedUse + "'");
        debug("");
    });
}

function testWithHMAC(usages, expectedUse)
{
    return crypto.subtle.importKey("raw", hmacKeyAsArrayBuffer, {name: 'hmac', hash: {name: 'sha-256'}}, extractable, usages).then(function(result) {
        return crypto.subtle.exportKey("jwk", result);
    }).then(function(result) {
        jwk = JSON.parse(bytesToASCIIString(result));
        debug(usages + ":");
        shouldBe("jwk.use", "'" + expectedUse + "'");
        debug("");
    });
}

Promise.all([
    testWithAESCBC(["encrypt"], "enconly"),
    testWithAESCBC(["decrypt"], "deconly"),
    testWithAESCBC(["encrypt", "decrypt"], "enconly,deconly"),
    testWithAESCBC(["wrapKey"], "wrap"),
    testWithAESCBC(["unwrapKey"], "unwrap"),
    testWithAESCBC(["wrapKey", "unwrapKey"], "wrap,unwrap"),
    testWithAESCBC(["encrypt", "decrypt", "wrapKey"], "enconly,deconly,wrap"),
    testWithAESCBC(["encrypt", "decrypt", "wrapKey", "unwrapKey"], "enc"),
    testWithHMAC(["sign"], "sigonly"),
    testWithHMAC(["verify"], "vfyonly"),
    testWithHMAC(["sign", "verify"], "sig"),
]).then(function() { finishJSTest(); } );
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
